# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Podcast Platform API",
        "description": "A comprehensive API for managing podcasts, episodes, and categories. This API provides endpoints for user authentication, podcast management, and content discovery.",
        "contact": {
            "name": "API Support",
            "url": "https://podcastplatform.com/support",
            "email": "support@podcastplatform.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/podcasts/{id}": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Get a specific podcast",
                "operationId": "a0df1edba7dc1534d35ad8be9995370f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of podcast",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podcast details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Podcast"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Podcast not found"
                    }
                }
            }
        },
...
        "/api/episodes/{id}": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Get a specific episode",
                "operationId": "ed03722f3d63a0e9c97eaa23c44c31ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of episode",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Episode details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Episode"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Episode not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Technology"
                    },
                    "slug": {
                        "type": "string",
                        "example": "technology"
                    },
                    "podcast_count": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "Podcast": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "The Tech Show"
                    },
                    "description": {
                        "type": "string",
                        "example": "A weekly podcast about technology and innovation"
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/images/tech-show.jpg"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "episode_count": {
                        "type": "integer",
                        "example": 50
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "Episode": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Episode 1: Introduction"
                    },
                    "description": {
                        "type": "string",
                        "example": "In this episode, we introduce the show and its hosts"
                    },
                    "audio_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/audio/ep1.mp3"
                    },
                    "duration": {
                        "description": "Duration in seconds",
                        "type": "integer",
                        "example": 3600
                    },
                    "podcast_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "podcast": {
                        "$ref": "#/components/schemas/Podcast"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use your API token as: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}